mysite/requirements.txt
Django==3.2.9
coverage==6.1.2
django-cors-headers==3.10.0
djangorestframework==3.12.4
psycopg2-binary==2.8.4
python-decouple==3.5
requests==2.26.0